+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.0.4] - 2025-02-26
+
+- fix #151 Allow the fromSerializationStrategy to handle null in place of an
+  serializable object.
+
 ## [v2.0.1 - v2.0.3] - 2022-04-15
 
 - NPM Deploy fixes